+++ checks/trunk-outdated-bindgen/src/main.rs
@@ -1,0 +1,4 @@
+fn main() {
+    console_log::init().unwrap();
+    log::debug!("Hi, this can run on the browser");
+}