+++ examples/trunk/flake.nix
@@ -85,6 +85,12 @@
         # This derivation is a directory you can put on a webserver.
         my-app = craneLib.buildTrunkPackage (commonArgs // {
           inherit cargoArtifacts;
+          # The version of wasm-bindgen-cli here must match the one from Cargo.lock.
+          wasm-bindgen-cli = pkgs.wasm-bindgen-cli.override {
+            version = "0.2.87";
+            hash = "sha256-0u9bl+FkXEK2b54n7/l9JOCtKo+pb42GF9E1EnAUQa0=";
+            cargoHash = "sha256-AsZBtE2qHJqQtuCt/wCAgOoxYMfvDh8IzBPAOkYSYko=";
+          };
         });
 
         # Quick example on how to serve the app,